+++ CHANGELOG.md
@@ -1,14 +1,16 @@
 ## Unreleased
 
 ### General
--
+- ノートを削除した際、関連するノートが同時に削除されないようになりました
+	- APIで、「replyIdが存在しているのにreplyがnull」や「renoteIdが存在しているのにrenoteがnull」であるという、今までにはなかったパターンが表れることになります
 
 ### Client
 - Fix: 一部の設定検索結果が存在しないパスになる問題を修正  
   (Cherry-picked from https://activitypub.software/TransFem-org/Sharkey/-/merge_requests/1171)
 
 ### Server
--
+- Enhance: ノートの削除処理の効率化
+- Enhance: 全体的なパフォーマンスの向上
 
 
 ## 2025.7.0